Rename module to GATK4
Reflects the intention better, this module will start by housing all
code related to the trial of GATK4.

Should be noted that some the code here may seem not directly related
to GATK4. The idea will be to refactor that code into a shared module
when the need arises (ie we add an ADAM experiment)