First experiment with GATK4
Adds a simple pipeline which uses a combination of Picard and GATK4.
- Convert an input FASTQ file into a uBAM
- Map the uBAM to a ref genome using the GATK BwaSparkEngine

uBAM is an unaligned BAM file
https://gatkforums.broadinstitute.org/gatk/discussion/11008/ubam-unmapped-bam-format
Should be noted that FastqToSam also sorts the file by queryname which
is required downstream.

The reference genome needs an index image file (also a GATK standard
thingy) which can be created from any .fa with the
BwaMemIndexImageCreator. Since these files are huge they are expected
to be provided externally to the codebase. May try to bake them into
a docker container later.

The pipeline is run within PipelineTest which now just trivially counts
the reads produced after mapping. Will add a more interesting functional
test later.