Create test to verify pipeline output
Modifies the test to compare the output of a pipeline run against
a pre-configured expected output (byte by byte). The expected output
can be found in /src/test/resources/expected. It comes from a test
pipeline run on crunch.

I'm not 100% sure its possible to compare BAM files this way (actually
doubt it) but 90% of the refactor can be used for another verification
mechanism.

In order to allow for this, needed to enable file output from the
pipeline. To do this we needed to be a bit more careful with using
the correct sequence dictionaries. The GATK provides a class that
writes the output BAM with spark, when given a good header and dict.

Added a simple convention for storing intermediate results files. They
are stored in a directory called "results" in the working directory of
the application.

Removed the tumour samples for now to simplify things.

The current test does not pass as I've used a different ref genome to
that which the expected bam was aligned.