Add reference genome to test resources
This commit adds a reference genome to the codebase to ease running
locally and on build agents (no need for an external reference).

We use the fa for chromosome 22 as its the smallest reference file. The
file is taken from
ftp://ftp.ensembl.org/pub/release-75/fasta/homo_sapiens/dna/. Will
probably split my own out of the illumina ref genome used in the
production pipeline as a next step.

Also small refactor to configuration to match hmf code style.