Verify BAM against output of old pipeline
Quite a few changes to make a real verification mechanism. First we
change the sample used for the whole pipeline. Instead of mapping a
full sample to a single chromosome reference, we map a slice of the
sample only for that chromosome. We're using chr 22 for everything as
it has the smallest reference file.

The reference files now use a slice from the same reference genome used
in production rather than the ones I originally found online

The pipeline now sorts the BAM by coordinate, as the old pipeline does.
Added this step when playing around with comparing BAM files. It's not
strictly necessary for the verification, but is a necessary step in
the pipeline so leaving in place.

The verification happens by reading both the expected (taken from a real
pipeline run) and the results using the htsjdk. We then read all
alignment records into two maps and compare read by read. We disregard
tags at this point as more work is required to mimic the tagging
exactly in GATK4 and doesn't really teach us anything. If all reads
are the equal and the size of the maps is equal, we consider the new
BAM verified.