Refactor to prepare for ADAM experiments
Up to this point the pipeline has been tightly coupled to GATK, as
that was the only framework under evaluation. Next we want to
try out ADAM, but keep as much of our test framework in place for
reuse.

To accomplish this we invert the dependencies in the pipeline and
introduce the concept of a stage. A stage declares its output and
contains a functional execute method. The pipeline has a map of stages
to expected outputs in the pipeline. The old pipeline is split up into
separate classes per stage.

This is far from a finished product, it is still quite tied to the
concepts in the GATK pipeline, has no validation or error handling/retry
and has implicit dependencies between stages in the map we'd like to
eventually make more declarative. There is also no way at this point
to parallelize two steps, which will likely be required later. But..
this will be enough to start with ADAM, which will inform the later
design.

At the module level, we create a new generic pipeline module with no
GATK dependencies. There we keep some of the conceptual and utility
classes from the old gatk4 module. GATK4 only contains classes with
direct gatk dependencies. The system module ties it all together, and
can later be used to mix and match any combination of stages.