First experiment with ADAM
This commit adds a new test to perform BWA and a coordinate sort using
the ADAM framework. ADAM provides a suite of APIs and models for
genomics research built to make use of Spark as an execution model and
platform.

The BWA portion makes use of the ADAM Pipe API (an extension to the
Spark PIPE API) to pipe an RDD of fragments into an externally running
bwa and then streams the outputs back into an RDD of alignment records
for saving or further processing. It requires us to use an interleaved
fastq input as we cannot pipe two files in serial. This as been included
as a prepared input (used SamToFastQ to create)

The coordinate sort is actually unnecessary as the sorting can simply
be done when persisting the aligned records in the BWA stage. Kept it
in for now for consistency with GATK pipeline.

Should be noted that if we go for ADAM we likely would not bother saving
every intermediate step as it could be much better for the Spark
optimizer to see the whole pipeline before choosing its plan.

Last note, the end sorted, aligned, BAMs are not identical. 2 flags are
different in the ADAM results from the old pipeline. Will dig further
but for now added a tolerance for missing reads of 1%.