Add lanes in preparation for merging stage
This commit adds the concepts of "lanes" (and a few other domain model
objects in the domain of sequencer output files). The stage and
pipeline interfaces change to be more functional.

Stage now accepts a single lane as an argument, rather than having the
configuration as a field in each object.

Pipeline accepts the whole sequence output, now it simply iterates
over the lanes, but in the future will handle the merge/mark dups
and eventually comparing real and tumour samples for variants.

The test now asserts that two lanes are processed. Result files are
expected to follow previous conventions with L00* in the file denoting
which lane.

These conventions are still too implicit throughout the code, so next
will do some refactoring to tidy that up.