Add mark duplicates (and merge) step GATK
Attempts to make use of the GATK4 markdups functionality. If I
understand correctly if we union RDDs from each lane and feed it
to the tool we should get a merged RDD, dups marked, back.

As of this commit the stage fails on an invalid input exception. I think
that is due to the fact my "lane 2" is just a straight copy of lane 1.
Going to do some more research before moving forward.

The merge is represented as a new stage that accepts a whole FlowCell
rather than a Lane.