Verifying mark duplicates against expected output
MarkDuplicatesSpark seems to have a bug where it cannot handle a BAM
with reads mapped multiple times, bombs out on a duplicate key
exception. To progress further with the verification I've moved to
using the Picard markdups for now and will revisit spark later.

Introduces samples from the hmf testdata 100k_reads_hiseq which are
really read from separate lanes (previously I'd just copied lane one
and called it lane 2). These samples are used in the new test of
merging and marking dups.

Started refactoring the assertions to be a bit tidyer as we add new
assertions for checking duplicates in two deduped files. Still needs
a bit more work to remove duplication.