Chain assertions on the same file
A bit of a quick and dirty refactor to allow for chaining file
assertions. The idea being you can declaritively test for both
correct alignment and markduplicates in the same test (rather than
running the pipeline twice).

The tests are still failing for the same reasons as before, so still
in the TODO to look at the functional validation in detail (but that
should also involve a good look at the samples we're using)