Decouple persistence from preprocessing
Refactor for single responsibility of the stages and allow for
different storage strategies without having to modify the pipeline
or stage itself. Also makes the stage interface more functional (one
input one output) which is generally easier to test and more composable.

Introduces Output, which has an entity (lane, sample,patient) and a
payload (rdd of data). Then the store interfaces are implemented for
each flavour.