Containerize pipeline
This change adds a profile to the build to enable running the pipeline
as a JAR and wraps that into a docker container. All dependencies are
copied into the /lib directory by the maven jar plugin.

The docker container is created with the spotify dockerfile maven
plugin, see its documentation for more detail.

The Dockerfile extends openjdk for no particular reason other than
most people do that for java applications. BWA is installed with
apt-get. A small script starts the jar and checks for error code.

The runtimes are combined in one class to make things simpler for the
user, we now choose GATK4 vs ADAM with a yaml property.

We do not yet push to dockerhub or private registry.