Push docker image to dockerhub via travis
This change will version all maven artifacts with the build number
(prefixed with 0.0. for semantic versioning) by running the maven
versions plugin as a before install step.

We also log into our hartwigmedicalfoundation dockerhub (password not
yet encrypted, as its still a dummy account really, we'll probably
use a private registry in the long run). Then the push is run as
a post install command. Did it this way rather than as part of the
deploy maven phase as we don't have a maven repository to deploy the
jars yet.