Remove need for interleaved fastq
In initial experiments I thought an interleaved fastq was required for
the pipe to bwa (I copied that from the cannoli project). After a bit
of research seems that it is also fine to load paired fastq as alignment
records (with no alignments) then run that through bwa. This alleviates
the need for creating interleaved fastq files.

Should note that the aligment still does not match the expectation,
but we'll look into that in more detail once functional validation
is done.