Conform to sequencer directory structure/naming
Up to this point we've been using a modified directory structure to
make things a little simpler for the tests, but now we'd like to run
against real samples so best to use the same structure as currently
used in production.

To accomplish this we can also make the code more robust by using the
real file names (see Lane readsFile and matesFile) rather than
reconstructing them later.

For now we only support the "R" and "T" convention for normal and
tumour samples. Single sample runs are no longer supported so we've
added a dummy tumour sample for the 100k test.

Enabled the ADAM test as a smoke test without its assertion for now.