Running pipeline2 in docker
Fixes a small issue when running in a docker container. The working dir
of the application is the root of the filesystem, so the results dir
ends up as //results/. This triggers a bug in the hadoop filesystem
resolver. So this change checks if we're at the root.

Also add some logging and a little timer to the pipeline. MS is probably
not the ideal unit but we'll enhance that later, I'm sure there will
be lots to time so we'll want a more sophisticated stopwatch.