Make running in docker a little easier
Added a default conf to the docker image that will use the cancer
panel sample for now, akin to the current run_pipeline.

Added a script to wrap the call to docker and make explicit the volumes
that need mounting to run the container. Will add this to the scripts
repo as well.