Fixing issue with ADAM BWA pairing
Found in production that when running BWA that pairs were not being
picked up. I think BWA was processing the pipe as a singed unpaired
reads file, and needs two arguments to switch to paired read mode.

To accomplish this we add the first reads as a file argument then
pipe the second reads in as before.