Split bwa and markduplicates for ADAM
In order to shorten the feedback loop in tuning mark duplicates,
persist the merged BWA result and check to see if the output of any
stage exists before (re)running it. Intentionally did not include a
force option for now as I think it's just as easy to clear the results
dir.