Consolidate logging to single file
BWA logs a lot to stderr, I think it would be cool to have everything
in one file, as pipeline2 is not very chatty. To accomplish this we
go back to only a console appender then redirect all output from the
start script. Also added a bit of trace to the pipeline to see what
stage is being run.