Validating result of Pipeline2 ADAM alignment
Having another good luck into automated validation in the functional
tests. After some experimenting and digging, a few small changes and
observations.

For regression sake we roll our bwa version back to 0.7.5a, which is
what the current pipeline uses. We also add a new expected BAM coming
from a run of the latest version of the pipeline.

We remove the "tolerance threshold" from the alignment assertion but we
keep a couple filters in place for pragmatism:

We do not validate unmapped reads. Since we only map to a ref genome
of chromosome 22, we're only interested in what is mapped and the unmapped
tend to be a bit unpredictable.

We also exclude duplicates from the alignment assertion as we will
validate those later.