Validate duplicate marking in ADAM
Took a deep dive into the workings of MarkDuplicates and how it
compares to sambamba. Two big things we learn:
- ADAM marks all secondary reads as duplicates
- Sambamba allows for duplicates with the mate unmapped, ADAM ignores
them.

So two changes:
- Remove the -M flag that marks many supplementary reads as secondary
- When validating filter out sambamba duplicates with MATE_UNMAPPED
(this is not the final state, need to discuss with team)

Also removed the optionality around persisting intermediate results and
just always persist them. Makes the code simpler and ensures we don't
run BWA multiple times if RDD is not cached. Also found this to be a
more stable practice in production, although it does add some runtime.