DEV-439 - Support single sample mode
Remove the concept of "RawSequenceOutput" in favour of a Reader and
factory pattern supporting two flavours of patient directory structure.

The first, single sample mode, which expects a directory full of fastq
files of a single sample matching that in the configuration exactly.

The second, two subdirectories suffixed with R and T. Each containing
the reference and tumour fastqs respectively.