DEV-441 Refactor configuration with YAML
Moves all code related to the filesystem into the system/runtime
packages, makes for a better dependency structure and is intuitive.

Remove the old configuration interface and use instead some immutable
value objects which look alot like the yaml file. Only tricky bit is
spark parameters are a map of strings, as we'd like to support any
spark property without having to list them in a massive interface.

JSonDeserialize and the yaml module are used to deserialize straight
into the value objects.