DEV-442 BWA thread configuration
With DEV-441 out of the way, much easier to add a configuration
parameter. That said, this is the first "optional" parameter, so
we introduce a pattern of using a @Default member of the value object.
This is preferable to using Optional for downstream clients of the class
and works a bit nicer with deserialization.