DEV-447 Explictly sort before saving
The boolean passed to "save" on an RDD does do any actual sorting,
just marks the appropriate flag in the header. Now we ensure sorting
in the sample store before saving.

A test is added that checks the reference index is in ascending order
and within each reference index the start alignment position is also
ascending.

While testing this noticed that BWA was being run twice, need to cache
the RDDs before union to ensure its only ever run once in a single
pipeline. The cast to AlignmentRecordRDD is necessary due to some
ugly scala integration.