DEV-446 Wire in indel realignment to pipeline
Previously we had a stub indel realignment stage that was not used. This
commit adds configuration for known indels which, when present will
enable the indel realignment stage.

The stage is run during the functional test but we still assert against
the mark duplicates BAM. ADAM indel realignment discards all the un-
mapped reads (ie non-chr22) and we don't have a proper validation
criteria yet for indel realignment.