DEV-447 Sort triggers multiple BWA runs
As both sortByReference and save trigger shuffle operations (sortby
and keyby) they end up triggering bwa twice. I'd tried to solve that
by caching the RDDs, but this didn't go well in production with the
size of the RDDs in memory.

This change moves sorting to after mark duplicates. This is required
as markduplicates does not preserve the sort order (would have been
nice to just have an explicit sort step and have that order preserved).