DEV-449 Germline calling with Avocado
Refactor the pipeline architecture to allow for a "germline calling"
stage that outputs not an AlignmentRecordRDD but a VariantContextRDD
which can be used to save a VCF containing variants from the germline.

This first commit is a naive calling of BiallelicGenotyper with default
parameters just to get a feel for it.

Germline calling is deemed an experimental feature so we add a toggle
and disable by default. It is also disabled in the test.

Already uncovered a version problem with avocado's spark version (2.2.0)
so need to downgrade ours to match.