DEV-449 Functioning germline calling
Moves to the latest version of Avocado so we can use their filtering
mechanism (which would be incompatible with the latest ADAM if we use
0.1.0).

Persists the input to BiallelicGenotyper. The warning will still show
as we do not persist to memory (I've found in practice it's too hard
on the heap), but should help performance nonetheless.

Biggest change is to use bwa 0.7.17 again, as 0.7.5a does not produce a
MD tag, which is used by avocado to find variants.

Also added a couple util methods to centralize suppression of redundant
casts (necessary due to the factoring of the scala code)