DEV-465 Remove GATK and unnecessary abstractions
Remove all references to GATK4 as we plan to focus on ADAM going forward

This means we can also remove the generics from the Pipeline, in favour
of explicit references to ADAM API classes. As a next step we could
really merge the pipeline and ADAM modules, but for now I'm keeping them
separate as it may still be nice to have a distinction (ie maybe
downstream clients depend on pipeline but don't want all the ADAM
specifics).

Also removed the generics used to allow for i/o of entities other than
samples.