DEV-461 Introduce QC checks for read counts
This change introduces a quality control check on each BAM enrichment
step (mark dups, realign indels, etc) to ensure no reads are dropped
along the way. As there was nothing like this before we also introduce
the concept of QC along with some refactoring.

QC steps are run on the inputs to each step to ensure we capture the
writing of the files. Eg. before we run indel realignment, we read in
mark duplicates and ensure its read count is the same as output of
BWA.

While testing this we discovered that ADAM indel realignment drops all
unmapped reads for performance. I have an open PR with them to fix this
for now our check only looks at the mapped read count.

To simplify things I've removed germline calling and renamed Pipeline
to BamCreationPipeline, as there is a pretty good chance we will have
separate pipelines for other parts of the whole analysis workflow.