DEV-462 Add final coverage QC check
Introduces a new QC to check coverage meets acceptance criteria.
Criteria is defined as the percentage of the genome that meets X times
coverage. EG. A threshold of 90% on 10x coverage means 90% of the ref
genome is aligned to at least 10 times.

ADAM provides a CoverageRDD which is a projection of the alignment RDD.
The downside of this approach is that we only get coverage metrics for
regions of the genome with at least one alignment. Investigation of a
better solution is ongoing, but this is why the ADAM context and ref
genome is unused in the current implementation.