DEV-461 Append unmapped reads to steps that filter them
Rather than waiting for ADAM to fix this (they are not sure they want
to) we'll do the fix on our side, and always append the unmapped to
after realign indels.

Also removing the AddMDTag step for now as we're not using germline
calling (and will likely switch to bwa 0.7.17).