DEV-465 Run for tumor sample in parallel
This change expands the bam creation pipeline to run over the the
tumor sample as well as the reference. This is done in parallel, so
the tumor sample kicks off at the same time as reference, but they
share a spark context.

The tumor run is optional, so if no tumor directory is found things will
continue to run in single sample mode.