DEV-474 Kubernetes/spark docker image
Includes a modified version of the spark Dockerfile which also includes
pipeline5 jars and configuration. The entire spark distribution is
added to version control for packaging. Perhaps a better way would be
to extend a spark image with our entry point, but let's leave that for
another commit.

Also switching to the fabric8 docker maven plugin as it allows for
multiple dockerfiles in the same build.