DEV-462 Use SAMRecord util to account for DELs in reads
Prior the handling of a DEL, which would introduce an alignment width
greater than the sequence of bases (qualities) in the read, which would
in turn cause an arrayindexoutofbounds, was handled by crudely trunc-
ating the base qualities.

This change uses SAMRecord static utilities, which uses the CIGAR, to
account for DELs in the sequence.

Currently the choice is made to not include the DEL in the coverage
count.