DEV-462 Do final QC even when no stages are run
Currently the final QC was only done when we actually run a stage but
that has proven a little inconvenient when validating the QC in prod
(need to wait hours before it starts).

This change will load the final stage for QC on every run, not sure
we'll keep this feature once QC is validated.