DEV-481 Massive refactor to allow for HDFS use
Had to remove all the assumptions about local filesystem. Simplified
things a bit in the process but also makes the application more
portable as a whole.

The hadoop FileSystem abstraction is used to easily switch between
running against a local filesystem or a distributed HDFS.The HDFS
location is configured in the yaml but will default local.

Also included a couple of k8 yaml files to quickly spin up an HDFS
cluster of a single namenode and datanode (TBC).

Had to start using the ADAM snapshot releases to move BWA from loading
a file directly from the filesystem due to a bug in the fragment
grouping we fixed recently.