DEV-481 Download reference genome and known indels from g storage
Another try to download static data before booting up the driver or
executor. Unfortunately the init container of spark is coupled to
downloading dependencies and not a try init container.

Instead just run the download script at the beginning of the entrypoint
which means also we need gcloud sdk available.