DEV-483 No patient name mode fixes
Quite a few problems when running without a patient name specified.
This should fix all of those and make the patient reading a little
more strict and intuitive. It's now expected that there will always
be a root directory of the patient name. The patient name yaml property
can be used to filter if there are multiple patients in the same
root.