DEV-500 DEV-501 Java bootstrap for Dataproc
Beginning of the bootstrap process for Google Dataproc using the Java
APIs over the bash scripted approach.

Adds the maven shade plugin to system build to create an uber jar
with all dependencies.

Adds a Dataproc cluster class which starts a cluster, submits a spark
job and stops the cluster, all using the Dataproc java sdk.