DEV-507 Run spark jobs in Dataproc via CLI tool
This commit completes a functional bootstrap which can start a
dataproc cluster, run the pipeline spark job, tear down the cluster
and clean up files.

The pipeline changes to distribute the reference genome to bwa via the
ADAM Pipe API rather than relying on it to be installed elsewhere.

Relocations are used in the shade plugin to avoid guava, etc dependency
conflicts. Settings taken from Google's own HDFS connector.

A main is introduced to bootstrap which can set all options for runtime.