DEV-507 Dockerize bootstrap
Creates a docker "binary" which can be used to run the bootstrap with
a fixed jar based on the current build. This binary will be versioned
along with the released pipeline.

The container is built in the patient-cluster module, which depends on
system. A test dependency is used to copy the jar over and ensure the
dependency graph is correct. That said, we don't want to have a compile
or runtime dependency on classes in system, as many are shaded.

Also disables the kuberenetes docker image build as it takes ages.
Keeping the supporting files for now in case our approach changes.