DEV-507 Move bootstrap to per sample and add basic S3 support
This change refactors the whole bootstrapping process to work
on a per sample basis rather than a per patient. This should improve
the throughput as we don't need to wait for both fastqs to land in
SBP S3. On cluster this equates to all samples running in "single
sample mode".

Also introduces streams to copy data from SBP S3 to Google Storage using
the AWS JDK.