DEV-507 Increase cluster size for GIAB
Going to a 6 node cluster. Seems that Dataproc will create one executor
per worker so we try to use as much of each worker as possible. A lot
of additional "yarn overhead" is added for BWA.

Also adds a no cluster delete options to leave the cluster running
when the job finishes.