DEV-519 Create a runtime bucket for each pipeline run
Adds the concept of a run and runtime bucket. Every interaction with
Google Storage happens via the runtime bucket to create the sandbox.

Also introduces a check for existence of the sample in the runtime
bucket to allow for re-using data already uploaded.

Static data is copied into the runtime bucket from well known buckets
to be exposed within the clusters staging bucket.

We also remove the restriction in patient reader on subdirectories to
allow for reuse of gunzipped fastqs.