DEV-523 Download final BAM to file or SBP object store
Adds a download step to the bootstrap with two flavours, local file
system or SBP object store. We create an input stream for the final
BAM and pass it to one or the other based on whether the sbp sample id
is set.

For SBP we create the BAM in a bucket of the form
hmf-bam-storage/${barcode}. The barcode is taken from the reads file
name. We also compute an MD5 for the sbp api, but this MD5 is not yet
updated in the object store.

The sbp api is called after the upload is complete, with the appropriate
metadata as a PATCH api call.