DEV-523 Fix issues with PATCH and http client
Java does not support PATCH as an http method out of the box, and fails
hard when you try. Jersey supports a workaround which we take advantage
of here.

This requires us to be a little more explicit in our dependencies,
ensuring we use a modern jersey-client and jax-rs api.