DEV-528 Index BAM at the end of bam creation
Calls the htsjdk BamIndexer at the end of the pipeline run to create
the BAI. The indexer is called in the driver, single threaded, using
the BAM file as input.  Tried using ADAM for this as well (converting
to SAMRecords RDD then indexing), but the indexer is not serializable
and I really doubt this is an operation that should be parallelized.

There is an open ticket to implement BAI generation in ADAM, probably
good we follow up on that one later if performance is a concern.