DEV-532 Use BWA SHM to share reference index across processes
This change runs bwa shm on each executor jvm to reduce the memory
consumption of bwa tasks. It is implemented using a JVM singleton
initialization task.

Unfortunately we need to add the files to the spark context twice to
ensure they are available for the initialization step as well as the
pipe of the AlignmentRecordRDD.

There is also a restriction on the length of the filename for shm, and
the Home_Sapiens_etc_etc exceeds this length. Therefore we alias all the
reference files with "reference", but only in the runtime bucket.