DEV-536 Use gsutil for upload
Adds a new sample upload that will use the gsutil cli for upload from
local filesystem or s3.

gsutil is run with a simple process builder. The path can be passed as
an argument or defaulted to the location in the docker container.

The Dockerfile now extends the cloud sdk parent and installs java
manually. This proved to be the easiest way to mix openjdk and gcloud
in my previous k8 experiments.

The old streaming upload is left in place until performance has been
proven.