DEV-538 Add a cluster optimizer to size cluster by input size
Refactors the notion of performance profiles to be generated from the
input files or S3 objects. It's nothing fancy, just looks at a linear
GB to CPU ratio (specified externally, default is 5) and creates the
number of 32 cpu boxes to support that.

The ratio is 5 by default but can also be overriden with the
cpu_gb_ratio option.

Also sets the spark executor memory to be 80% of the available memory
minus the yarn overhead, this is a hard restriction of dataproc.