DEV-538 Scale down executor cores
Total memory used at any time is task size * cores. Given the default
level of parallelism, using all the cores eats up too much memory. This
is just to see the effect, there are better ways we can fix this and
not waste the CPU (ie highmem boxes or repartitioning)