DEV-541 Add monitor and stackdriver implementation
Adds the simple concepts of metrics and a monitor and creates a
stackdriver implementation. The stackdriver APIs are quite straight-
forward. You create a descriptor for your metric and then create it
as a timeseries.

A service account is required to access the monitor from within the
cluster GCEs. This is specified in the cluster config and not currently
configurable via the bootstrap.

Run id and version are passed to the spark job as args and used as
labels for later filtering. Only time spent aligning and markduping/
realigning, and total bam creating are now logged.