DEV-545 User google billing api to calculate cost
Gets the true SKU cost from google's billing api for the compute
engine. Dataproc cost is calculated with a fixed constant 0.01 per cpu
cost as no SKU is available in the API. No discounts are applied yet.

Also wires in the monitor to the bootstrap (using the bootstrap
credentials) and logs end run cost as a new metric.