DEV-558 Create status file to pass along to bootstrap
Introduces a status reporter which will create a file in hadoop marking
the run a _SUCCESS or _FAILURE. The latter file will contain the reason
for failure.

Some refactoring in the pipeline to not throw runtime exceptions on QC
failures.

In the bootstrap we check for the success or fail file in the results
bucket. This is passed along to the SBP api if appropriate, or just
logged if we're running in local mode.