DEV-550 Back to explicit vcores
When turning off minimum vcores and the dominant resource calculator
we've observed some odd behaviour. Halfway throw the realign indels
stage we loose 75% of our executors and things slow down.

More investigation required, but as we know this does not occur with
these settings, using them for now. Small adjustment, we use 80% of the
32 CPUs in the hope we can allocate 7 containers (1 driver and 6
workers)