DEV-560 Refactor BAM creation to remove all intermediate persistence
Simplifies the BAM creation a lot by only supporting to fixed and
explicit stages: bwa/alignment; enrichment. The latter is a combo of
mark dups and indel realignment. This split still exists so we can do
the read count check on the enrichment. No results are persisted until
the final BAM.

We also remove support for reference/tumour here as the idea will be to
move that logic into the bootstrap and that BAM creation doesn't need
to know what type of sample its running on.