DEV-560 Fixing memory issues in indel realignment
As per my chat with the ADAM developers, using a Kryo registrator to
reduce the impact of Kryo serialization on memory. Using a registrator
will keep track of classes by an int index rather than the full class
name. Unfortunately could not use the ADAM provided registrator, as it
caused a stackoverflow exception.