DEV-564 Add composer to bootstrap and remove sorting
First we modify the pipeline to write out the BAM as head and tail (ie
a with headerless shards) and defer merging. Also removing sort at this
point as we'll do this soon externally, so will make testing this
feature a bit quicker.

In the bootstrap we create a composer which will concatenate all the
tail parts to the BAM header. Google restricts compose operations to
32 components, so we need to compose them into 32 component partitions
then recursively reduce them.

Also disabled the old docker build, not really used so just taking time
and space at this point. Can easily be turned back on later.