DEV-566 Use Sambamba (or samtools) for indexing and sorting
Refactor the indexer to spawn sambamba to sort and index the bam. The
full BAM is copied from hdfs (google storage) to the local working dir
then sambamba is run with as many threads as processors on the box.
The result is copied back to hdfs/gs.

If sambamba is not available will fall back to samtools. This is handy
for running locally on a Mac, where installing sambamba is not
possible.

The whole thing is run as a separate pipeline in its own cluster, this
way we can have a single node working on it and we don't waste any
money on a big idle spark cluster.

The node init script is modified to install sambamba.