DEV-570 Use transcoding to reduce gunzipping time
First step is to set the object metadata on the samples properly in the
bootstrap. This will tell Google Storage APIs to decompress on these
objects when requested.

On the requester (pipeline) we change the gunzipper to only rename the
files when in Google Cloud runtime. This is so Spark doesn't try to
unzip them on newHadoopApi calls.