DEV-613 Add retry mechanism to dataproc jobs
As per the JIRA, this is added explicitly for failures due to pre-
emptible VMs (although unfortunately we have no explicit way to test
for that exact failure scenario). So if a job fails remotely, we
assume this is the most likely cause and we try once more.

In preparation, refactored the duplication in Bootstrap into a simple
Job framework where each job starts a cluster, submits a job, stops
the cluster and tracks the metrics. It also checks the status of the
job, and does the retry if fails. Note, checking the status is
currently only available for BAM creation. It also retries if any
exception is thrown by the cluster operation, in the case of vms
failure this would occur in the submit method.