Get barcode from SBP rest api
Turns out we were actually parsing out the flowcell id and not the
barcode from the name of the fastq. SBP provides the sample API to
extract the barcode properly. This change adds this extra api call
and wires it through.

Also moved the SBP specific stuff to its own package for more clarity
into what classes are used for this code path.