DEV-631 Register shutdown hook which deletes clusters/buckets
This is to better handle the situation where a SIGTERM is sent when the
node on which the bootstrap is running is taken down for maintenance,
or other planned circumstance.

Should be noted that the JVM does not make a distinction between a
normal shutdown and one triggered by an external signal.

There is still a redundant finally block which will still run before
the shutdown block on normal execution. Leaving for now until I see a
bit more how this behaves in the wild.