DEV-636 Add simple janitor process to find and cleanup stuff
The janitor will be a simple daemon container which at a specified
interval (5 mins for now) will execute a list of cleanup tasks.

To start we only have one task: find clusters running with no jobs
scheduled or running. This will cover the big risk of expensive clusters
dangling after a bootstrap crash.

The janitor is released as a container, and will run indefinitely once
started. It can be restarted with impunity.