DEV-641 Re-attach to running spark jobs
Checks to see if the spark job which is about to be run is already
running to prevent duplicate work. Jobs are now assigned a natural
id (default was to randomly generate).

If a job exists but it not running (cancelled, done, error, etc), the
job will be deleted and re-submitted.