DEV-641 Port back to 25.0 and resume from crashed bam create
In parallel runs one run crashed in bam creation on a spark job failing.
The bootstrap cleanup up the runtime bucket (so logs are lost) along
with the unzipped samples.

However, the gunzip job was marked complete so when K8 restarted the
failed job we began from bam creation, with zipped fastqs. This is not
supported and we got in a failure loop.

This change removes the deletion of the runtime bucket on errors, which
is clever anyway for forensics. But to make sure that corrupt results
are not left over for re-runs, the bam pipeline now will clean its
target directory before starting.

Also had to port back to 0.25.0 as the ADAM guys have made a bit of a
mess of 0.26.0-SNAPSHOT.