Fix issue with multiple transfers and base64 encoded md5
Fixes bug where the transfer was always running twice. Also adds test
for this feature.

Encodes the base64 hash from google into hex, which is what SBP expects.

Resolves some nasty classpath issues that cropped up pinning 0.25.0.
Something started going wrong in the shading, so disabled for now for
guava (we're all on 20). But this will likely need to be added again
later.

Remove retries on individual jobs. K8 retries will cover this and it
was causing issues with keeping track of job logs.