DEV-650 Enable rclone for upload and download.
Adds an option to the bootstrap to use rclone over gsutil and the AWS
Java API for transfering data between google storage, S3 and the local
filesystem. Also provided are options to customize the GS and S3 rclone
remotes, and the path the rclone.

Also refactored all copy operations to sit behind an simple interface,
made it easier to unit test.