DEV-650 Fix race condition on shared process builder
Last commit we removed the parallel processing of fastq upload assuming
there was an issue running multiple rclones concurrently. This
assumption appears to be false (cannot reproduce outside of bootstrap).

The problem seems to actually be a shared process builder between
threads. This change removes the shared stateful process builder and
uses a supplier instead to create a new one for each file, but stil
allow for easy unit testing.